feat: separate claw and arm
